requirements.txt
pycryptodome
cryptography
pandas
matplotlib
numpy